use devenv from devenv if in devenv ;)

+++ shell/ci/testing/setup-devenv.sh
@@ -44,16 +44,24 @@ if [[ -n $CI ]]; then
   if ! command -v devenv >/dev/null; then
     info "Setting up devenv"
 
-    tempDir=$(mktemp -d)
-    cp "$DIR/../../../.tool-versions" "$tempDir/" # Use the versions from devbase
-    pushd "$tempDir" >/dev/null || exit 1
-    gh release -R getoutreach/devenv download --pattern "devenv_*_$(go env GOOS)_$(go env GOARCH).tar.gz"
-    echo "" # Fixes issues with output being corrupted in CI
-    tar xf devenv**.tar.gz
-    sudo mv devenv /usr/local/bin/devenv
-    sudo chown circleci:circleci /usr/local/bin/devenv
-    rm -rf "$tempDir"
-    popd >/dev/null || exit
+    # If we're in the devenv repo, use a local build.
+    if [[ "$(yq -r '.name' service.yaml)" == "devenv" ]]; then
+      info_sub "Using local devenv build"
+      make
+      info_sub "built binary: $(./bin/devenv version)"
+      sudo cp ./bin/devenv /usr/local/bin/devenv
+    else
+      tempDir=$(mktemp -d)
+      cp "$DIR/../../../.tool-versions" "$tempDir/" # Use the versions from devbase
+      pushd "$tempDir" >/dev/null || exit 1
+      gh release -R getoutreach/devenv download --pattern "devenv_*_$(go env GOOS)_$(go env GOARCH).tar.gz"
+      echo "" # Fixes issues with output being corrupted in CI
+      tar xf devenv**.tar.gz
+      sudo mv devenv /usr/local/bin/devenv
+      sudo chown circleci:circleci /usr/local/bin/devenv
+      rm -rf "$tempDir"
+      popd >/dev/null || exit
+    fi
   fi
 
   info "Setting up Git"
